Update to the latest runes protocol (#15)
* Update to the latest rune protocol

* Update local.sh

* fix tests

---------

+++ customs/bitcoin/src/lib.rs
@@ -1134,7 +1134,7 @@ pub fn build_unsigned_transaction(
             .map(|(idx, (_, amount))| Edict {
                 id: rune_id.into(),
                 amount: *amount,
-                output: (idx + 2) as u128,
+                output: (idx + 2) as u32,
             })
             .collect::<Vec<Edict>>(),
     };
